admin users list view and edit view and process

+++ java/com/springapp/blog/controllers/admin/CategoryController.java
@@ -1,6 +1,7 @@
-package com.springapp.blog.controllers;
+package com.springapp.blog.controllers.admin;
 
 import com.springapp.blog.dao.CategoryDao;
+import com.springapp.blog.entity.Category;
 import com.springapp.blog.models.CategoryModel;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -8,6 +9,8 @@ import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.validation.FieldError;
 import org.springframework.web.bind.annotation.*;
+
+import javax.validation.Valid;
 
 @Controller
 @RequestMapping(value = "/admin/categories")
@@ -37,7 +40,7 @@ public class CategoryController {
     }
 
     @PostMapping("/create")
-    public String createProcess(Model model, @ModelAttribute("category") CategoryModel categoryModel, BindingResult bindingResult){
+    public String createProcess(Model model, @ModelAttribute("category") @Valid CategoryModel categoryModel, BindingResult bindingResult){
         if (bindingResult.hasErrors()){
             model.addAttribute("view", "admin/categories/create");
             return "base-layout";
@@ -56,14 +59,26 @@ public class CategoryController {
 
     @GetMapping("/edit/{id}")
     public String edit(Model model, @PathVariable("id") Integer id){
+        Category category = categoryDao.getOne(id);
+
+        if (category == null){
+            return "redirect:/admin/categories/";
+        }
+
         model.addAttribute("view", "admin/categories/edit");
-        model.addAttribute("category", categoryDao.getOne(id));
+        model.addAttribute("category", category);
 
         return "base-layout";
     }
 
     @PostMapping("/edit/{id}")
-    public String editProcess(Model model, @PathVariable("id") Integer id, @ModelAttribute("category") CategoryModel categoryModel, BindingResult bindingResult){
+    public String editProcess(Model model, @PathVariable("id") Integer id, @ModelAttribute("category") @Valid CategoryModel categoryModel, BindingResult bindingResult){
+        Category category = categoryDao.getOne(id);
+
+        if (category == null){
+            return "redirect:/admin/categories/";
+        }
+
         if (bindingResult.hasErrors()){
             model.addAttribute("view", "admin/categories/edit");
             return "base-layout";
@@ -76,6 +91,12 @@ public class CategoryController {
 
     @GetMapping("/delete/{id}")
     public String delete(Model model, @PathVariable("id") Integer id){
+        Category category = categoryDao.getOne(id);
+
+        if (category == null){
+            return "redirect:/admin/categories/";
+        }
+
         model.addAttribute("view", "admin/categories/delete");
         model.addAttribute("category", categoryDao.getOne(id));
 
@@ -84,8 +105,15 @@ public class CategoryController {
 
     @PostMapping("/delete/{id}")
     public String deleteProcess(@PathVariable("id") Integer id){
+        Category category = categoryDao.getOne(id);
+
+        if (category == null){
+            return "redirect:/admin/categories/";
+        }
+
         categoryDao.delete(id);
 
         return "redirect:/admin/categories/";
     }
+
 }
